Extract shared symbolic printers
Move residual and measurement code generation into a shared `_symbolic_printers` package, and update the compiled model plus tests to use the new import surface. Adds coverage for the shared printers, measurement layout normalization, and bicomplex/cfunc wiring.
